Refactor label inspecion and intention to support and reuse intentions also for environments
