+++ resources/intentionDescriptions/LatexAddLabelToCommandIntention/after.tex.template
