+++ resources/intentionDescriptions/LatexAddLabelToCommandIntention/before.tex.template
