+++ resources/intentionDescriptions/LatexAddLabelToCommandIntention/description.html
