+++ resources/intentionDescriptions/LatexAddLabelToEnvironmentIntention/after.tex.template
@@ -1,0 +1,2 @@
+\begin{lstlisting}[label={lst:listing}]
+\end{lstlisting}
