+++ resources/intentionDescriptions/LatexAddLabelToEnvironmentIntention/before.tex.template
@@ -1,0 +1,2 @@
+\begin{lstlisting}
+\end{lstlisting}
