+++ resources/intentionDescriptions/LatexAddLabelToEnvironmentIntention/description.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+Add a label.
+</body>
+</html>
